% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{compare.toplines}
\alias{compare.toplines}
\title{Make a dot and ribbon plot comparing symmetrical questions}
\usage{
compare.toplines(varlist, vartext = NULL, mulaw, sortResponse,
  remove = NULL)
}
\arguments{
\item{varlist}{a vector of quoted variable names}

\item{vartext}{a character vector of text which should be used to describe the variables}

\item{mulaw}{the relevant integ file}

\item{sortResponse}{The response category in the topline tables by which to sort the vartext}

\item{remove}{an optional vector of values to remove from the topline table, supplied to the remove
argument in make.topline()}
}
\value{
A ggplot object.
}
\description{
\code{compare.toplines} creates a scatter plot with time on the x-axis
}
\details{
This function takes a list of variables, makes toplines of them, and plots the toplines in a graph with
 one question per row and dots for the values of each question response.
}
\examples{
compare.toplines(varlist = c("d38", "d55", "d91", "d59", "d92"),
mulaw = demprimary,
vartext = c("Joe Biden", "Elizabeth Warren", "Kamala Harris", "Bernie Sanders", "Pete Buttigieg"),
sortResponse = "Favorable",
remove = c("don't know", "refused"))
}
