% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCrosstab.R
\name{crosstab}
\alias{crosstab}
\title{Make a weighted crosstab of row proportions}
\usage{
crosstab(mulaw, x, y, weight = zwave_weight, remove = "", n = TRUE,
  pct_type = "row", format = "wide")
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{logical, if TRUE a column of row totals is included}

\item{pct_type}{Controls the kind of percentage values returned. One of "row," "cell," or "column."}

\item{format}{one of "long" or "wide"}
}
\value{
A tibble
}
\description{
\code{crosstab} returns a data.frame containing a weighted crosstab of two variables
}
\details{
This is a wrapper around pollster::crosstab()
}
\examples{
crosstab(mulaw = df, x = zpid3, y = g40v2, remove = c("don't know", "refused"))
}
