% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCrosstab3Way.R
\name{crosstab_3way}
\alias{crosstab_3way}
\title{Make a weighted crosstab of row proportions, by wave}
\usage{
crosstab_3way(mulaw, x, y, z, weight = zwave_weight, remove = c(""),
  n = TRUE, pct_type = "row", format = "wide")
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{z}{The third variable}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{Logical. If TRUE add a row total column}

\item{pct_type}{one of "row" or "cell"}

\item{format}{one of "wide" or "long}
}
\value{
A tibble
}
\description{
\code{crosstab_3way} returns a data.frame containing a weighted crosstab of two variables
by wave.
}
\details{
This is a wrapper around pollster::crosstab_3way()
}
\examples{
crosstab_3way(integ, zpid3, g40, ac7)

}
