% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePalettes.R
\name{fav4.palette}
\alias{fav4.palette}
\title{\code{fav4.palette} returns a vector of colors for 4-category generic positive/negative}
\usage{
fav4.palette(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, reverses the order of the directionally relevant colors, but not the trailing colors}
}
\value{
A character vector
}
\description{
\code{fav4.palette} returns a vector of colors for 4-category generic positive/negative
}
