% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCrosstabWave.R
\name{make.crosstab.wave}
\alias{make.crosstab.wave}
\title{Make a weighted crosstab of row proportions, by wave}
\usage{
make.crosstab.wave(x, y, mulaw, remove, date = zpolldatestr,
  weight = zwave_weight)
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{date}{The date variable, defaults to zpolldatestr}

\item{weight}{The weighting variable, defaults to zwave_weight}
}
\value{
A dataframe.
}
\description{
\code{make.crosstab.wave} returns a data.frame containing a weighted crosstab of two variables
by wave.
}
\details{
This function returns a crosstab by wave. The resulting data.frame is in long format, suitable
 for making a faceted plot. See the example below
}
\examples{
make.crosstab.wave(x = zpid3, y = g40v2, mulaw = df, remove = c("don't know", "refused"))

# Now make that output a faceted graph
make.crosstab.wave(x = zpid3, y = g40v2,
                   mulaw = df, remove = c("don't know", "refused")) \%>\%
     ggplot(aes(zpid3, ae2, col = ae2)) +
       geom_point() +
       geom_smoth() +
       facet_wrap(facets = vars(zpid3))
}
