% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTopline.R
\name{make.topline}
\alias{make.topline}
\title{Make a time series of a variable}
\usage{
make.topline(variable, mulaw, weight = zwave_weight, n = TRUE,
  cumsum = TRUE, remove)
}
\arguments{
\item{variable}{The variable, unquoted}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{weight}{The weight variable, defaults to zwave_weight}

\item{n}{Logical, determines if a row total is included or not}

\item{cumsum}{Logical, determines if a cumulative percent column is included or not}

\item{remove}{an optional vector of response options to remove from the table (after calculating percents).
This is useful when making tables.}
}
\value{
A dataframe.
}
\description{
\code{make.topline} returns a data.frame containing a time series of a variable
}
\details{
This function returns a topline table of any given variable as a data.frame.
}
\examples{
make.topline(variable = g2, mulaw = orig, remove = c("refused"))

}
