% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTimeSeries.R
\name{make.ts}
\alias{make.ts}
\title{Make a time series of a variable}
\usage{
make.ts(variable, mulaw, remove, format)
}
\arguments{
\item{variable}{The variable, unquoted}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).}

\item{format}{One of "wide" or "long", defaults to "wide"}
}
\value{
A dataframe.
}
\description{
\code{make.ts} returns a data.frame containing a time series of a variable
}
\details{
This function returns a time series of any given variable. The resulting data.frame defaults
 to wide format, but can be in long format if format = "long" is specified.
}
\examples{
make.ts(variable = g2, mulaw = orig, remove = c("refused"), format = "long")

}
