% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTimeSeries.R
\name{make_ts}
\alias{make_ts}
\title{Make a time series of a variable}
\usage{
make_ts(mulaw, variable, date = zpollenddate, weight = zwave_weight,
  remove = "", n = TRUE, pct_type = "row", format = "wide")
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{variable}{The variable}

\item{date}{the date variable, defaults to zpollenddate}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{logical, if TRUE a column of row totals is included}

\item{pct_type}{Controls the kind of percentage values returned. One of "row," "cell," or "column."}

\item{format}{one of "long" or "wide"}
}
\value{
A tibble
}
\description{
\code{make.ts} returns a data.frame containing a time series of a variable
}
\details{
This is a wrapper around pollster::crosstab in which x is assumed to be zpollenddate.
 All you have to provide is the y-var to the `variable` argument
}
\examples{
make_ts(integ, g40)

}
