% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPCT.R
\name{mlsp}
\alias{mlsp}
\title{Return a rounded marginal}
\usage{
mlsp(mulaw, variable, response, weight = zwave_weight)
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{variable}{The unquoted variable name}

\item{response}{The desired response. It can be either numeric or the value label.}

\item{weight}{The weighting variable, defaults to zwave_weight}
}
\value{
A dataframe.
}
\description{
\code{mlsp} returns an integer value for a marginal. It's the percent of valid respondents
giving the specified response.
}
