% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{mlspCrosstabBar}
\alias{mlspCrosstabBar}
\title{Make a grouped bar graph of a crosstab table}
\usage{
mlspCrosstabBar(crosstabtable, titlevar = NULL, title = NULL,
  subtitle = NULL, theme = "default", LubarSlides = TRUE,
  PlotMargins = c(0.25, 0, 2, 1), wraptitle = 90,
  legendPosition = "top", legendJust = "right", xlabelAngle = NULL,
  xlabelWrap = 12, facetncol = NULL)
}
\arguments{
\item{titlevar}{A quoted variable name. If supplied, the full text of this variable will be the title}

\item{title}{A character string. If supplied it will be the graph's title.}

\item{subtitle}{If supplied, it will be the graph's subtitle}

\item{theme}{Optional. One of either "LubarSlides" or "MLSP". This sets the theme to
theme_LubarSlides or theme_MLSP, respectively.}

\item{PlotMargins}{Only applicable when theme = "LubarSlides". It allows the user to manually change the plot.margin parameter.}

\item{wraptitle}{= the length at which to wrap the character title string. By default
it is 90 which matches the width of a LubarSlide.}

\item{legendPosition}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{legendJust}{anchor point for positioning legend inside plot ("center" or two-element numeric vector) or the justification according to the plot area when positioned outside the plot}

\item{xlabelAngle}{optional, the angle of the x-axis labels}

\item{xlabelWrap}{character length at which to wrap x-axis labels}

\item{facetncol}{optional, the number of facet columns}

\item{toplinetable}{The source table created by make.crosstab()}
}
\value{
A dataframe.
}
\description{
\code{mlspCrosstabBar} creates a bar graph of a crosstab table
}
\details{
This function takes a crosstab table created with make.crosstab() and returns it as
 a bar chart.
}
