% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{mlspTimeSeriesScatter}
\alias{mlspTimeSeriesScatter}
\title{Make a time series scatterplot}
\usage{
mlspTimeSeriesScatter(timeseriestable, titlevar = NULL, title = NULL,
  subtitle = NULL, theme = "default", LubarSlides = TRUE,
  PlotMargins = c(0.25, 0, 2, 1), wraptitle = 95,
  legendPosition = "top", legendJust = "right")
}
\arguments{
\item{timeseriestable}{The source table created by make.ts()}

\item{titlevar}{A quoted variable name. If supplied, the full text of this variable will be the title}

\item{title}{A character string. If supplied it will be the graph's title.}

\item{subtitle}{If supplied, it will be the graph's subtitle}

\item{theme}{Optional. One of either "LubarSlides" or "MLSP". This sets the theme to
theme_LubarSlides or theme_MLSP, respectively.}

\item{PlotMargins}{Only applicable when theme = "LubarSlides". It allows the user to manually change the plot.margin parameter.}

\item{wraptitle}{= the length at which to wrap the character title string. By default
it is 95 which matches the width of a LubarSlide.}
}
\value{
A dataframe.
}
\description{
\code{mlspTimeSeriesScatter} creates a scatter plot with time on the x-axis
}
\details{
This function takes a topline table created with make.topline() and returns it as
 a bar chart.
}
\examples{
make.ts(g40, integ) \%>\% mlspTimeSeriesScatter()
}
