% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{mlspToplineBar}
\alias{mlspToplineBar}
\title{Make a simple bar graph of a topline table}
\usage{
mlspToplineBar(toplinetable, titlevar = NULL, title = NULL,
  subtitle = NULL, theme = NULL, PlotMargins = c(0.25, 0, 2, 1),
  wraptitle = 90, xlabelAngle = NULL, xlabelWrap = 12)
}
\arguments{
\item{toplinetable}{The source table created by make.topline()}

\item{titlevar}{A quoted variable name. If supplied, the full text of this variable will be the title}

\item{title}{A character string. If supplied it will be the graph's title.}

\item{subtitle}{If supplied, it will be the graph's subtitle}

\item{theme}{Optional. One of either "LubarSlides" or "MLSP". This sets the theme to
theme_LubarSlides or theme_MLSP, respectively.}

\item{PlotMargins}{Only applicable when theme = "LubarSlides". It allows the user to manually change the plot.margin parameter.}

\item{wraptitle}{= the length at which to wrap the character title string. By default
it is 90 which matches the width of a LubarSlide.}

\item{xlabelAngle}{optional, the angle of the x-axis labels}

\item{xlabelWrap}{character length at which to wrap x-axis labels}

\item{xlab}{an optional x-axis name}
}
\value{
A dataframe.
}
\description{
\code{mlspToplineBar} creates a bar graph of a topline table
}
\details{
This function takes a topline table created with make.topline() and returns it as
 a bar chart.
}
\examples{
make.topline(variable = g2, mulaw = orig, remove = c("refused"), format = "long")
}
