% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeCrosstab.R
\name{moe.crosstab}
\alias{moe.crosstab}
\title{Make a time series of a variable}
\usage{
moe.crosstab(x, y, mulaw, weights = zwave_weight, format = "long")
}
\arguments{
\item{x}{The grouping variable}

\item{y}{the dependent variable}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{weights}{The weight variable, defaults to zwave_weight}

\item{format}{One of "wide" or "long", defaults to "wide"}
}
\value{
A dataframe.
}
\description{
\code{moe.crosstab} returns a data.frame a crosstab of two variables along with margins of error
}
\details{
This function returns a crosstab of two supplied variables along with margins of error.
 The resulting data.frame defaults
 to long format, but can be in wide format if format = "wide" is specified.
}
\examples{
moe.crosstab(zpid3, g40, integ, format = "wide")

}
