% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeCrosstabWave.R
\name{moe.crosstab.wave}
\alias{moe.crosstab.wave}
\title{Make a time series of a variable}
\usage{
moe.crosstab.wave(x, y, mulaw, date = zpollenddate,
  weights = zwave_weight, format = "long")
}
\arguments{
\item{x}{The grouping variable}

\item{y}{the dependent variable}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{date}{the date variable. Defaults to zpollenddate}

\item{weights}{The weight variable, defaults to zwave_weight}

\item{format}{One of "wide" or "long", defaults to "long"}
}
\value{
A dataframe.
}
\description{
\code{moe.crosstab.wave} returns a data.frame of a crosstab of two variables by wave along with margins of error
}
\details{
This function returns a crosstab  by waveof two supplied variables along with margins of error.
 The resulting data.frame defaults
 to long format, but can be in wide format if format = "wide" is specified.
}
\examples{
moe.crosstab.wave(zpid3, g40, integ, format = "wide")

}
