% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeCrosstab3Way.R
\name{moe.crosstab3way}
\alias{moe.crosstab3way}
\title{Make a 3-way crosstab of a variable}
\usage{
moe.crosstab3way(x, y, z, mulaw, weights = zwave_weight,
  format = "long")
}
\arguments{
\item{x}{The grouping variable}

\item{y}{the dependent variable}

\item{z}{the 2nd control variable.}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{weights}{The weight variable, defaults to zwave_weight}

\item{format}{One of "wide" or "long", defaults to "long"}
}
\value{
A dataframe.
}
\description{
\code{moe.crosstab3way} returns a data.frame of a crosstab of two variables by a 3rd variable
 along with margins of error
}
\details{
This function returns a crosstab by a 3rd variable of two supplied variables along with margins of error.
 The resulting data.frame defaults
 to long format, but can be in wide format if format = "wide" is specified.
}
\examples{
moe.crosstab.wave(zpid3, g40, integ, format = "wide")

}
