% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeTopline.R
\name{moe.topline}
\alias{moe.topline}
\title{Make a topline of a variable}
\usage{
moe.topline(x, mulaw, weights = zwave_weight)
}
\arguments{
\item{x}{The variable, unquoted}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{weights}{The weight variable, defaults to zwave_weight}
}
\value{
A dataframe.
}
\description{
\code{moe.topline} returns a data.frame containing a topline table with the margin of error
}
\details{
This function returns the topline statistics of any given variable.
}
\examples{
make.ts(variable = g2, mulaw = orig, remove = c("refused"), format = "long")

}
