% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeToplineWave.R
\name{moe.topline.wave}
\alias{moe.topline.wave}
\title{Make a topline by wave of a variable}
\usage{
moe.topline.wave(x, mulaw, date = zpollenddate, weights = zwave_weight,
  format = "long")
}
\arguments{
\item{x}{The grouping variable}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{date}{the date variable, defaults to zpollenddate}

\item{weights}{The weight variable, defaults to zwave_weight}

\item{format}{One of "wide" or "long", defaults to "wide"}
}
\value{
A dataframe.
}
\description{
\code{moe.topline.wave} returns a data.frame a crosstab of two variables along with margins of error
}
\details{
This function returns a crosstab of two supplied variables along with margins of error.
 The resulting data.frame defaults
 to long format, but can be in wide format if format = "wide" is specified.
}
\examples{
moe.crosstab(zpid3, g40, integ, format = "wide")

}
