% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeCrosstab.R
\name{moe_crosstab}
\alias{moe_crosstab}
\title{Make a crosstab of a variable}
\usage{
moe_crosstab(
  mulaw,
  x,
  y,
  weight = zwave_weight,
  remove = "",
  n = TRUE,
  pct_type = "row",
  format = "wide",
  zscore = 1.96,
  ...
)
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{logical, if TRUE a column of row totals is included}

\item{pct_type}{Controls the kind of percentage values returned. One of "row" or "cell." Column percents are not supported.}

\item{format}{one of "long" or "wide"}

\item{zscore}{defaults to 1.96, consistent with a 95\% confidence interval}

\item{...}{further arguments passed to pollster::moe_crosstab, such as unwt_n}
}
\value{
A tibble
}
\description{
\code{moe_crosstab} returns a data.frame a crosstab of two variables along with margins of error
}
\details{
this is a wrapper around pollster::moe_crosstab
}
\examples{
moe_crosstab(integ, zpid3, g40)
}
