% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeCrosstabWave.R
\name{moe_crosstab_wave}
\alias{moe_crosstab_wave}
\title{Make a crosstab by wave of a variable}
\usage{
moe_crosstab_wave(
  mulaw,
  x,
  y,
  z = zpollenddate,
  weight = zwave_weight,
  remove = c(""),
  n = TRUE,
  pct_type = "row",
  format = "long",
  zscore = 1.96,
  ...
)
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{z}{The third variable, defaults to zpollenddate}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{Logical. If TRUE add a row total column}

\item{pct_type}{one of "row" or "cell"}

\item{format}{one of "wide" or "long}

\item{zscore}{defaults to 1.96, consistent with a 95\% confidence interval}

\item{...}{further arguments passed to pollster::moe_crosstab_3way, such as unwt_n}
}
\value{
A tibble
}
\description{
\code{moe_crosstab_wave} returns a data.frame of a crosstab of two variables by wave along with margins of error
}
\details{
this is a wrapper around pollster::moe_crosstab_wave where z defaults to zpollenddate
}
\examples{
moe_crosstab_wave(integ, zpid3, g40)
}
