% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeToplineWave.R
\name{moe_time_series}
\alias{moe_time_series}
\title{Make a topline by wave of a variable}
\usage{
moe_time_series(mulaw, variable, date = zpollenddate,
  weight = zwave_weight, remove = "", n = TRUE, pct_type = "row",
  format = "long", zscore = 1.96)
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{variable}{The variable}

\item{date}{the date variable, defaults to zpollenddate}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{logical, if TRUE a column of row totals is included}

\item{pct_type}{Controls the kind of percentage values returned. One of "row," "cell," or "column."}

\item{format}{one of "long" or "wide"}

\item{zscore}{defaults to 1.96, consistent with a 95\% confidence interval}
}
\value{
A tibble
}
\description{
\code{moe.topline.wave} returns a data.frame a crosstab of two variables along with margins of error
}
\details{
This is a wrapper around pollster::moe_crosstab in which x is assumed to be zpollenddate.
 All you have to provide is the y-var to the `variable` argument
}
\examples{
moe_time_series(integ, g40)
}
