% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeTopline.R
\name{moe_topline}
\alias{moe_topline}
\title{Make a topline of a variable}
\usage{
moe_topline(mulaw, variable, weight = zwave_weight, remove = c(""),
  n = TRUE, pct = TRUE, valid_pct = TRUE, cum_pct = TRUE,
  zscore = 1.96)
}
\arguments{
\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{variable}{The variable, unquoted}

\item{weight}{The weight variable, defaults to zwave_weight}

\item{remove}{a character string of response values to remove from the output after performing all calculations}

\item{n}{Logical, determines if a row total is included or not}

\item{pct}{logical, determines if the pct column is included or not}

\item{valid_pct}{logical, determines if the valid percent column is included}

\item{cum_pct}{Logical, determines if a cumulative percent column is included or not}

\item{zscore}{defaults to 1.96, consistent with a 95% confidence interval}
}
\value{
A tibble
}
\description{
\code{moe_topline} returns a data.frame containing a topline table with the margin of error
}
\details{
This is a wrapper around pollster::moe_topline()
}
\examples{
moe_topline(integ, g40)

}
