% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePalettes.R
\name{pid3.palette}
\alias{pid3.palette}
\title{\code{pid3.palette} returns a vector of colors for 3-category party ID}
\usage{
pid3.palette(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, reverses the order of the directionally relevant colors, but not the trailing colors}
}
\value{
A character vector
}
\description{
\code{pid3.palette} returns a vector of colors for 3-category party ID
}
