% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePalettes.R
\name{pid5.palette}
\alias{pid5.palette}
\title{\code{pid5.palette} returns a vector of colors for 5-category party ID}
\usage{
pid5.palette(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, reverses the order of the directionally relevant colors, but not the trailing colors}
}
\value{
A character vector
}
\description{
\code{pid5.palette} returns a vector of colors for 5-category party ID
}
