% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.word.R
\name{print.word}
\alias{print.word}
\title{Print a number as an english word}
\usage{
\method{print}{word}(number, case)
}
\arguments{
\item{number}{any numeric value}

\item{case}{either "upper" or "lower." Defaults to "upper."}
}
\value{
A character string
}
\description{
\code{print.word} a character string of a number
}
\details{
This function is a wrapper for either the words() or Words() function from
 the english package. It defaults to upper case, but lower case can
 be set using case = lower. print.word() also adds a hyphen between words.
}
\examples{
print.word(42)
print.word(42, case = "lower")
}
