% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundingFunctions.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Forces rounded numbers to total their UNrounded sum, to any power of 10.}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{a vector of numeric values}

\item{digits}{what to round to, defaults to 0}
}
\value{
A character string
}
\description{
\code{round_preserve_sum}
}
\details{
This functions returns a rounded vector of numbers, ensuring that they add
 to the same total as before rounding.
}
\examples{
round_preserve_sum(c(2.2,4.5,8.7))
round_preserve_sum(c(2.2,4.5,8.7), digits = 1)
}
