% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePalettes.R
\name{sequence3.palette}
\alias{sequence3.palette}
\title{\code{sequence3.palette} returns a vector of colors for 3-category unidirectional sequence}
\usage{
sequence3.palette(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, reverses the order of the directionally relevant colors, but not the trailing colors}
}
\value{
A character vector
}
\description{
\code{sequence3.palette} returns a vector of colors for 3-category unidirectional sequence
}
