% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePalettes.R
\name{sequence4.palette}
\alias{sequence4.palette}
\title{\code{sequence4.palette} returns a vector of colors for 4-category unidirectional sequence}
\usage{
sequence4.palette(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, reverses the order of the directionally relevant colors, but not the trailing colors}
}
\value{
A character vector
}
\description{
\code{sequence4.palette} returns a vector of colors for 4-category unidirectional sequence
}
