% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotThemes.R
\name{theme_LubarSlides}
\alias{theme_LubarSlides}
\title{theme for tables with scaling and font sizes optimized for Lubar Center slideshows}
\usage{
theme_LubarSlides(PlotMargins = c(0.25, 0, 2, 1))
}
\arguments{
\item{PlotMargins}{A vector of margins for the graph in the form c(top, right, bottom, left). Units are lines.}
}
\description{
theme for tables with scaling and font sizes optimized for Lubar Center slideshows
}
