% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePalettes.R
\name{vote.palette}
\alias{vote.palette}
\title{\code{vote.palette} returns a vector of colors for Democrat vs Republican}
\usage{
vote.palette(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, reverses the order of the directionally relevant colors, but not the trailing colors}
}
\value{
A character vector
}
\description{
\code{vote.palette} returns a vector of colors for Democrat vs Republican
}
