% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCrosstab.R
\name{make.crosstab}
\alias{make.crosstab}
\title{Make a weighted crosstab of row proportions}
\usage{
make.crosstab(x, y, mulaw, remove, weight = zwave_weight, n = TRUE)
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{mulaw}{The data.frame containing the version of the integrated file you wish to use}

\item{remove}{An optional character vector of values to remove from final table (e.g. DK/Ref).
This will not affect any calculations made. The vector is not case-sensitive.}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{n}{logical, if TRUE a column of row totals is included}
}
\value{
A dataframe.
}
\description{
\code{make.crosstab} returns a data.frame containing a weighted crosstab of two variables
}
\details{
The x variable will appear as rows in the data.frame. The y variable will be columns.
}
\examples{
make.crosstab(x = zpid3, y = g40v2, mulaw = df, remove = c("don't know", "refused"))
}
